Uplifty Interivew Stephen Chan
Backend including bonuses implemented.

Design Doc https://xd.adobe.com/view/245a11e6-bf01-40f2-b737-493dbe0de6c1-7109/

Insomnia Graphql Testing Endpoints:

https://drive.google.com/file/d/1leyPnzUPZ73j9QpSGZGWlNkpVrzOvwi1/view?usp=sharing

Notes: I decided to focus on shaping the api endpoints and wanted to implement redux to manage the responses, but I did not get around to implementing it.

Backend work met, according to design specifications:

1. Standard deck -> Done
2. Deal 5 cards -> Done
3. 5 cards are random -> Done, using session info
4. Card Counter -> Done
5. Reset button -> Done
6. When all dealt, game over -> Done
7. Ace victory/loss -> Done
8. Unit Tests -> Partial
9. Bonuses have been implemented

I have included a design doc that i wanted to reflect the appropriate endpoints i was shaping, and had intended to carry out their implementation using redux to manage the state and map them to components appropirately.

The interview user loaded with the project is now tied to fixtures i have also loaded to the project, which are cards and decks. These objects are managed by the endpoints and history is tracked with a "game" model.

To use the insomnia endpoint workspace, you need to download insomnia. It is similar to postman.

A detailed explanation explaining design decisions can be found in the adobe XD document.

As always, i am available to talk about this project, and I wanted to say that I enjoyed myself doing it.

Again, thanks for the opportunity!